refactor(quran): stop leaking the PageAyahRange DAO type into upper layers
AP-1: the Room @Query projection PageAyahRange (declared in QuranDao) was
imported by domain (QuranRepository, QuranUseCases) and presentation
(QuranViewModel, QuranPageGrid). Add a PageAyahRange domain model, rename
the Room projection to PageAyahRangeRow, and map row -> domain in
QuranRepositoryImpl. domain/ now imports nothing from data/, and the two
presentation PageAyahRange leaks (AP-2) are cleared.

Verified with compileDebugKotlin and the full unit-test suite.
